SMILE-1268: mayor merge from main
