+++ create-publication-from-scopus/src/main/java/no/sikt/nva/scopus/ScopusConstants.java
@@ -1,5 +1,6 @@
 package no.sikt.nva.scopus;
 
+import no.unit.nva.model.contexttypes.PublicationContext;
 import nva.commons.core.JacocoGenerated;
 
 public class ScopusConstants {
@@ -9,6 +10,12 @@ public class ScopusConstants {
 
     //URI constants:
     public static final String DOI_OPEN_URL_FORMAT = "https://doi.org";
+
+    //Hournal constants:
+    public static final String ISSN_TYPE_ELECTRONIC = "electronic";
+    public static final String ISSN_TYPE_PRINT = "print";
+    public static final PublicationContext EMPTY_PUBLICATION_CONTEXT = null;
+    public static final String ERROR_MSG_ISSN_NOT_FOUND = "Could not find issn";
 
     @JacocoGenerated
     public ScopusConstants() {
