+++ create-publication-from-scopus/src/main/java/no/sikt/nva/scopus/ScopusConverter.java
@@ -1,6 +1,9 @@
 package no.sikt.nva.scopus;
 
 import static no.sikt.nva.scopus.ScopusConstants.DOI_OPEN_URL_FORMAT;
+import static no.sikt.nva.scopus.ScopusSourceType.JOURNAL;
+import static nva.commons.core.attempt.Try.attempt;
+
 import jakarta.xml.bind.JAXB;
 import java.io.StringWriter;
 import java.net.URI;
@@ -13,7 +16,9 @@ import no.scopus.generated.AuthorGroupTp;
 import no.scopus.generated.AuthorTp;
 import no.scopus.generated.CollaborationTp;
 import no.scopus.generated.DocTp;
+import no.scopus.generated.IssnTp;
 import no.scopus.generated.ItemidTp;
+import no.scopus.generated.SourceTp;
 import no.scopus.generated.TitletextTp;
 import no.scopus.generated.YesnoAtt;
 import no.unit.nva.metadata.CreatePublicationRequest;
@@ -22,11 +27,19 @@ import no.unit.nva.model.Contributor;
 import no.unit.nva.model.EntityDescription;
 import no.unit.nva.model.Identity;
 import no.unit.nva.model.Reference;
+import no.unit.nva.model.contexttypes.PublicationContext;
+import no.unit.nva.model.contexttypes.UnconfirmedJournal;
+import no.unit.nva.model.exceptions.InvalidIssnException;
+import nva.commons.core.SingletonCollector;
 import nva.commons.core.paths.UriWrapper;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 class ScopusConverter {
 
+    public static final String DASH = "-";
     private final DocTp docTp;
+    private static final Logger logger = LoggerFactory.getLogger(ScopusConverter.class);
 
     protected ScopusConverter(DocTp docTp) {
         this.docTp = docTp;
@@ -64,6 +77,7 @@ class ScopusConverter {
     private Reference generateReference() {
         Reference reference = new Reference();
         reference.setDoi(extractDOI());
+        reference.setPublicationContext(getPublicationContext());
         return reference;
     }
 
@@ -162,5 +176,63 @@ class ScopusConverter {
     private AdditionalIdentifier toAdditionalIdentifier(ItemidTp itemIdTp) {
         return new AdditionalIdentifier(ScopusConstants.ADDITIONAL_IDENTIFIERS_SCOPUS_ID_SOURCE_NAME,
                                         itemIdTp.getValue());
+    }
+
+    private PublicationContext getPublicationContext() {
+        if (isJournal(docTp)) {
+            return attempt(() -> createUnconfirmedJournal(docTp))
+                    .orElseThrow(fail -> logErrorAndThrowException(fail.getException()));
+        }
+        return ScopusConstants.EMPTY_PUBLICATION_CONTEXT;
+    }
+
+    private RuntimeException logErrorAndThrowException(Exception exception) {
+        logger.error(exception.getMessage());
+        return exception instanceof RuntimeException
+                ? (RuntimeException) exception
+                : new RuntimeException(exception);
+    }
+
+    private UnconfirmedJournal createUnconfirmedJournal(DocTp docTp) throws InvalidIssnException {
+        var source = getSource(docTp);
+        var sourceTitle = extractSourceTitle(source);
+        var issnTpList = source.getIssn();
+        var printIssn = findPrintIssn(issnTpList).orElse(null);
+        var electronicIssn = findElectronicIssn(issnTpList).orElse(null);
+        return new UnconfirmedJournal(sourceTitle, printIssn, electronicIssn);
+    }
+
+    private boolean isJournal(DocTp docTp) {
+        return ScopusSourceType.valueOfCode(docTp.getMeta().getSrctype()) == JOURNAL;
+    }
+
+    private SourceTp getSource(DocTp docTp) {
+        return docTp.getItem().getItem().getBibrecord().getHead().getSource();
+    }
+
+    private String extractSourceTitle(SourceTp sourceTp) {
+        StringBuilder sourceTitle = new StringBuilder();
+        sourceTp.getSourcetitle().getContent().forEach(sourceTitle::append);
+        return sourceTitle.toString();
+    }
+
+    private Optional<String> findElectronicIssn(List<IssnTp> issnTpList) {
+        return findIssn(issnTpList, ScopusConstants.ISSN_TYPE_ELECTRONIC);
+    }
+
+    private Optional<String> findPrintIssn(List<IssnTp> issnTpList) {
+        return findIssn(issnTpList, ScopusConstants.ISSN_TYPE_PRINT);
+    }
+
+    private Optional<String> findIssn(List<IssnTp> issnTpList, String issnType) {
+        return Optional.ofNullable(issnTpList.stream()
+                .filter(issn -> issnType.equals(issn.getType()))
+                .map(IssnTp::getContent)
+                .map(this::addDashToIssn)
+                .collect(SingletonCollector.collectOrElse(null)));
+    }
+
+    private String addDashToIssn(String issn) {
+        return issn.contains(DASH) ? issn : issn.substring(0, 4) + DASH + issn.substring(4);
     }
 }
